[Gap Decorations]: Parse row-rule-visibility-items
This CL parses the `row-rule-visibility-items` property, which controls
how gap decorations interact in segments with empty cells. Although this
property is not yet part of the official spec, we are introducing it
behind the Gap Decorations flag, as this appears to be the intended
direction for the feature. Any changes following discussions with the
working group will be updated accordingly.

See: https://github.com/w3c/csswg-drafts/issues/12602

Bug: 357648037
Change-Id: I102e64ecfeb161f31600d931c1718bf37cb5d9ea